services/lighton-ocr/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
httpx==0.27.0
pillow==10.4.0
transformers==4.43.3
